+++ app/Http/Controllers/MilestonesController.php
@@ -10,6 +10,7 @@ use App\Http\Requests\Milestones\CreateMilestoneRequest;
 use App\Http\Requests\Milestones\ReorderMilestonesRequest;
 use App\Http\Requests\Milestones\UpdateMilestoneRequest;
 use App\Models\Milestone;
+use App\Models\User;
 use App\Services\Items\MilestoneItemService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
@@ -44,7 +45,7 @@ class MilestonesController extends BaseController
      */
     public function index(Request $request)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $args = $request->only(['page', 'per_page', 'sort', 'order_by', 'search']);
 
@@ -64,7 +65,7 @@ class MilestonesController extends BaseController
      */
     public function show(Request $request, Milestone $milestone)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         return Inertia::render('Domains/Partials/ManageMilestone', [
             'milestone' => $milestone,
@@ -77,7 +78,7 @@ class MilestonesController extends BaseController
      */
     public function store(CreateMilestoneRequest $request)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $attributes = $request->validated();
         $result     = $this->milestoneItemService->create($attributes);
@@ -94,7 +95,7 @@ class MilestonesController extends BaseController
      */
     public function update(UpdateMilestoneRequest $request, Milestone $milestone)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $attributes = $request->validated();
         $result     = $this->milestoneItemService->update($milestone->id, $attributes);
@@ -111,7 +112,7 @@ class MilestonesController extends BaseController
      */
     public function destroy(Request $request, Milestone $milestone)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $result = $this->milestoneItemService->delete($milestone->id);
 
@@ -127,7 +128,7 @@ class MilestonesController extends BaseController
      */
     public function restore(Request $request, Milestone $milestone)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $result = $this->milestoneItemService->update($milestone->id, [
             'deleted_at' => null,
@@ -144,7 +145,7 @@ class MilestonesController extends BaseController
      */
     public function reorder(ReorderMilestonesRequest $request)
     {
-//        $this->authorize('authorizeAdminAccess', User::class);
+        $this->authorize('authorizeAdminAccess', User::class);
 
         $attributes = $request->validated();
         $result     = $this->milestoneItemService->reorder($attributes);
